pom trouble cause
