+++ src/main/java/com/afs/tas/exceptions/IVVRuntimeException.java
@@ -1,0 +1,8 @@
+package com.afs.tas.exceptions;
+
+public class IVVRuntimeException extends RuntimeException {
+
+    public IVVRuntimeException(String message){
+        super(message);
+    }
+}
