--- src/main/java/gov/dhs/tsa/tas/exceptions/IVVRuntimeException.java
@@ -1,8 +1,0 @@
-package gov.dhs.tsa.tas.exceptions;
-
-public class IVVRuntimeException extends RuntimeException {
-
-    public IVVRuntimeException(String message){
-        super(message);
-    }
-}
