--- src/main/java/gov/dhs/tsa/tas/exceptions/PendingException.java
@@ -1,12 +1,0 @@
-package gov.dhs.tsa.tas.exceptions;
-
-public class PendingException extends RuntimeException {
-
-    public PendingException(String message) {
-        super(message);
-    }
-
-    public PendingException(){
-        super("This step not yet implemented.");
-    }
-}
