+++ src/test/java/com/afs/tas/exceptions/PendingExceptionTest.java
@@ -1,0 +1,17 @@
+package com.afs.tas.exceptions;
+
+import org.testng.annotations.Test;
+
+
+public class PendingExceptionTest {
+
+    @Test(expectedExceptions = PendingException.class)
+    public void constructorTest() {
+        throw new PendingException("This is the exception.");
+    }
+
+    @Test(expectedExceptions = PendingException.class)
+    public void defaultConstructorTest(){
+        throw new PendingException();
+    }
+}
