--- src/test/java/gov/dhs/tsa/tas/exceptions/IVVRuntimeExceptionTest.java
@@ -1,12 +1,0 @@
-package gov.dhs.tsa.tas.exceptions;
-
-import org.testng.annotations.Test;
-
-
-public class IVVRuntimeExceptionTest {
-
-    @Test(expectedExceptions = RuntimeException.class)
-    public void testIVVRuntimeExceptionTest(){
-        throw new IVVRuntimeException("This message");
-    }
-}
