--- src/test/java/gov/dhs/tsa/tas/exceptions/PendingExceptionTest.java
@@ -1,17 +1,0 @@
-package gov.dhs.tsa.tas.exceptions;
-
-import org.testng.annotations.Test;
-
-
-public class PendingExceptionTest {
-
-    @Test(expectedExceptions = PendingException.class)
-    public void constructorTest() {
-        throw new PendingException("This is the exception.");
-    }
-
-    @Test(expectedExceptions = PendingException.class)
-    public void defaultConstructorTest(){
-        throw new PendingException();
-    }
-}
