+++ src/test/resources/testng.everything.xml
@@ -10,8 +10,8 @@
         </run>
     </groups>
     <listeners>
-        <listener class-name="gov.dhs.tsa.tas.listeners.SuiteListener"/>
-        <listener class-name="gov.dhs.tsa.tas.listeners.TestListener"/>
+        <listener class-name="com.afs.tas.listeners.SuiteListener"/>
+        <listener class-name="com.afs.tas.listeners.TestListener"/>
     </listeners>
     <test name="Everything">
         <packages>
