+++ src/test/resources/testng.unit.xml
@@ -11,8 +11,8 @@
         </run>
     </groups>
     <listeners>
-        <listener class-name="gov.dhs.tsa.tas.listeners.SuiteListener"/>
-        <listener class-name="gov.dhs.tsa.tas.listeners.TestListener"/>
+        <listener class-name="com.afs.tas.listeners.SuiteListener"/>
+        <listener class-name="com.afs.tas.listeners.TestListener"/>
     </listeners>
     <test name="Regression">
         <packages>
